pyproject.toml
[project]
name = "semstash"
version = "0.1.0"
description = "Unlimited semantic storage for humans and AI agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
keywords = ["semantic", "storage", "embeddings", "vector", "s3", "mcp", "ai", "multimodal"]
authors = [
    { name = "Danilo Poccia" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "boto3>=1.35.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "mcp>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    # Web UI templates
    "jinja2>=3.1.0",
    # Document support
    "pymupdf>=1.25.0",  # PDF to image rendering
    "python-docx>=1.1.0",  # DOCX text extraction
    "python-pptx>=1.0.0",  # PPTX text extraction
    # Spreadsheet support (XLSX to CSV/text)
    "openpyxl>=3.1.0",
    # Document to Markdown conversion
    "markitdown[pdf,docx,pptx,xlsx]>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "moto[s3]>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "boto3-stubs[s3,bedrock-runtime,pricing]>=1.35.0",
    "types-openpyxl>=3.1.0",
    "pdoc>=15.0.0",
    "vulture>=2.14",
    "pre-commit>=4.0.0",
    # Strands Agent SDK for MCP server testing
    "strands-agents>=0.1.0",
    # Security scanning
    "bandit>=1.7.0",
    "pip-audit>=2.7.0",
]

[project.scripts]
semstash = "semstash.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/semstash"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.ruff.lint.per-file-ignores]
# Ignore N803 (camelCase args) in conftest.py for AWS SDK mock functions
"tests/conftest.py" = ["N803"]
# Ignore B008 (function call in default) in web.py for FastAPI Query() defaults
"src/semstash/web.py" = ["B008"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -m 'not integration'"
asyncio_mode = "auto"
markers = [
    "integration: requires real AWS credentials and resources (use --use-aws to run)",
]

[tool.coverage.run]
source = ["src/semstash"]
omit = ["*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
